Move Pyramid class to rdd.py
